Changed how server handles and responds to upvote/downvote
userVote is still +1/-1 when sent to the server (to indicate upvote or
downvote) but is sent *back* as the users current standing with the
link, whether that be +1, 0, or -1. This allows a user to upvote a link,
then press the downvote button and return to a neutral state (and vice
versa). Coincidentally, this ended up simplying the front end display
functions as well. The links new score is also now sent back as a
response to upvote/downvote.